Use python `logger` instead of `print`
By default, the imported `logging` object is an instantied
logger (called `ROOT`). So we only need to configure it
once at the beginning of the main module, then we can import
this configured instance everywhere in the project - without
the need to configure it again or import the instance from a
local module!

Alternatively, we can import and configure the root logger,
then instanciate other Logger but these will automatically
inherit the configuration we did in the ROOT logger.

That's a great example of a pattern built into the language!